✨ feat: profile 코드 수정 및 중첩 레이아웃 적용 / 오타 수정 / 프로필 => 각 메뉴로 라우팅 되도록 폴더 생성
